# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interactive API for testing Neuron keylib functionality",
        "title": "Keylib API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/signatures/components": {
            "post": {
                "description": "Parse a signature and return its R, S, V components",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signatures"
                ],
                "summary": "Get signature components",
                "parameters": [
                    {
                        "description": "Signature hex",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ParseSignatureRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.SignatureComponentsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/signatures/parse": {
            "post": {
                "description": "Parse and validate a signature from hex string (65 bytes: R || S || V)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signatures"
                ],
                "summary": "Parse a signature from hex",
                "parameters": [
                    {
                        "description": "Signature hex",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ParseSignatureRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.SignatureComponentsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/signatures/recover-from-digest": {
            "post": {
                "description": "Recover the public key that signed a 32-byte digest from the signature",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signatures"
                ],
                "summary": "Recover public key from signed digest",
                "parameters": [
                    {
                        "description": "Recover from digest request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.RecoverFromDigestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.PublicKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/signatures/recover-pubkey": {
            "post": {
                "description": "Recover the public key that signed a message from the signature",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signatures"
                ],
                "summary": "Recover public key from signed message",
                "parameters": [
                    {
                        "description": "Recover request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.RecoverPublicKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.PublicKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid hex format"
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ParseSignatureRequest": {
            "type": "object",
            "required": [
                "signatureHex"
            ],
            "properties": {
                "signatureHex": {
                    "type": "string",
                    "example": "0x...65bytes..."
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.PublicKeyResponse": {
            "type": "object",
            "properties": {
                "evmAddress": {
                    "type": "string",
                    "example": "0xe364f2f1e5f4f03d1df682322500b9c68c997ec3"
                },
                "evmAddressChecksum": {
                    "type": "string",
                    "example": "0xe364f2f1e5F4F03d1df682322500b9c68C997ec3"
                },
                "peerID": {
                    "type": "string",
                    "example": "16Uiu2HAm3Lkn9NRieuh3UUTWMNthSDumQL9ctTBKxQqdCC79WUSq"
                },
                "publicKeyHex": {
                    "type": "string",
                    "example": "0x02759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae"
                },
                "publicKeyHexUncompressed": {
                    "type": "string",
                    "example": "0x04759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae..."
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.RecoverFromDigestRequest": {
            "type": "object",
            "required": [
                "digestHex",
                "signatureHex"
            ],
            "properties": {
                "digestHex": {
                    "type": "string",
                    "example": "0x1234567890abcdef...32bytes..."
                },
                "signatureHex": {
                    "type": "string",
                    "example": "0x...65bytes..."
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.RecoverPublicKeyRequest": {
            "type": "object",
            "required": [
                "message",
                "signatureHex"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello, World!"
                },
                "signatureHex": {
                    "type": "string",
                    "example": "0x...65bytes..."
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.SignatureComponentsResponse": {
            "type": "object",
            "properties": {
                "isZero": {
                    "type": "boolean",
                    "example": false
                },
                "r": {
                    "type": "string",
                    "example": "0x..."
                },
                "s": {
                    "type": "string",
                    "example": "0x..."
                },
                "v": {
                    "type": "integer",
                    "example": 0
                },
                "vEthereum": {
                    "type": "integer",
                    "example": 27
                }
            }
        },
...
    }
}